pyproject.toml
[tool.poetry]
name = "github-actions-with-r"
version = "0.1.0"
description = ""
authors = ["y-mattsuda <m.youichi1225@icloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
slack-sdk = "^3.16.2"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
black = "^22.3.0"
flake8 = "^4.0.1"
python-dotenv = "^0.20.0"
mypy = "^0.960"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py39']
line-length = 119

[tool.isort]
profile = "black"
line_length = 119

[tool.mypy]
python_version = "3.9"
no_strict_optional = true
ignore_missing_imports = true
check_untyped_defs = true
